requirements.txt
discord
colorama
asyncio
python-dotenv
pytz
requests
aiohttp
psutil
openai
PyYAML
piper-tts
twitchio
python-osc
simpleaudio